Break cyclic dependencies by moving BaseUnitData subclasses in one module
