+++ module/data/_module.mjs
@@ -3,6 +3,7 @@ export {default as SystemDataModel, SparseDataModel, getLocalizeKey} from "./abs
 export * as actor from "./actor/_module.mjs";
 export * as advancement from "./advancement/_module.mjs";
 export * as chat from "./chat/_module.mjs";
+export * as common from "./common/_module.mjs";
 export * as fields from "./fields/_module.mjs";
 export * as item from "./item/_module.mjs";
 export * as roll from "./roll/_module.mjs";
